chg: usr: major refactor: plot is method of annealer.
