+++ adannealing/__init__.py
@@ -1,5 +1,4 @@
 from .annealer import Annealer
-from .plotting import plot
 try:
     from ._version import __version__
 except ImportError:
